requirements.txt
absl-py==2.1.0
actionlib==1.14.0
aiohttp==3.9.5
aiosignal==1.3.1
angles==1.9.13
async-timeout==4.0.3
attrs==23.2.0
bondpy==1.8.6
cachetools==5.4.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
Cython==3.0.10
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
docstring_parser==0.16
dynamic-reconfigure==1.7.3
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.1
h5py==3.10.0
idna==3.7
image-geometry==1.16.2
importlib_metadata==8.1.0
importlib_resources==6.4.0
interactive-markers==1.12.0
Jinja2==3.1.4
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonargparse==4.32.0
kiwisolver==1.4.5
laser_geometry==1.6.7
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
message-filters==1.16.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.1
pillow==10.3.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytorch-lightning==2.3.3
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
topic-tools==1.16.0
torch==2.2.2
torchmetrics==1.4.0.post0
torchvision==0.17.2
tqdm==4.66.4
triton==2.2.0
typeshed_client==2.7.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
xacro==1.14.17
yarl==1.9.4
zipp==3.19.2
